Added character expansions.
CoolType will now skip over strings found in the expansions array and will not try to type them out character by character.

This allows you to use things like &nbsp; in strings passed to coolType. As long as &nbsp; is found in the expansions array (which it is by default) then it will insert &nbsp; as one character instead of 6.